# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sakura Chronicles MMORPG API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ServerStatus {
  online: boolean;
  players: number;
  maxPlayers?: number;
  ping: number;
  uptime: string;
  region: string;
  lastUpdated: string;
}

export interface NewsArticle {
  id: number;
  title: string;
  summary: string;
  content: string;
  author: string;
  publishedAt: string;
  category: string;
  imageUrl: string;
  featured?: boolean;
}

export interface GameRelease {
  id: number;
  version: string;
  releasedAt: string;
  isLatest: boolean;
  changelog: string[];
  downloadUrlAndroid?: string;
  downloadUrlWindows?: string;
  downloadUrlIos?: string;
  sizeAndroid?: string;
  sizeWindows?: string;
  sizeIos?: string;
}

export type WikiEntryAttributes = {[key: string]: string};

export interface WikiEntry {
  id: number;
  slug: string;
  name: string;
  category: string;
  description: string;
  imageUrl?: string;
  attributes?: WikiEntryAttributes;
  lore?: string;
}

export interface RegisterInput {
  username: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RecoverInput {
  email: string;
}

export interface AuthResult {
  token: string;
  username: string;
}

export interface MessageResult {
  message: string;
}

export type ListNewsParams = {
/**
 * Max results to return
 */
limit?: number;
/**
 * Filter by category
 */
category?: string;
};