# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sakura Chronicles MMORPG API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: server-status
    description: Game server status
  - name: news
    description: News articles
  - name: downloads
    description: Game downloads and versions
  - name: wiki
    description: Wiki entries
  - name: auth
    description: Authentication

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /server-status:
    get:
      operationId: getServerStatus
      tags: [server-status]
      summary: Get game server status
      responses:
        "200":
          description: Server status information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerStatus"

  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List all news articles
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: Max results to return
        - in: query
          name: category
          schema:
            type: string
          description: Filter by category
      responses:
        "200":
          description: List of news articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"

  /news/{id}:
    get:
      operationId: getNewsArticle
      tags: [news]
      summary: Get a single news article
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found

  /downloads:
    get:
      operationId: listDownloads
      tags: [downloads]
      summary: List available downloads and version history
      responses:
        "200":
          description: List of releases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameRelease"

  /downloads/latest:
    get:
      operationId: getLatestRelease
      tags: [downloads]
      summary: Get the latest game release
      responses:
        "200":
          description: Latest release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameRelease"

  /wiki/{category}:
    get:
      operationId: listWikiEntries
      tags: [wiki]
      summary: List wiki entries by category
      parameters:
        - in: path
          name: category
          required: true
          schema:
            type: string
            enum: [classes, weapons, professions, bosses, dungeons]
      responses:
        "200":
          description: Wiki entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WikiEntry"

  /wiki/{category}/{slug}:
    get:
      operationId: getWikiEntry
      tags: [wiki]
      summary: Get a single wiki entry
      parameters:
        - in: path
          name: category
          required: true
          schema:
            type: string
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wiki entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WikiEntry"
        "404":
          description: Not found

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login to an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials

  /auth/recover:
    post:
      operationId: requestPasswordRecovery
      tags: [auth]
      summary: Request password recovery email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoverInput"
      responses:
        "200":
          description: Recovery email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ServerStatus:
      type: object
      required: [online, players, ping, uptime, region, lastUpdated]
      properties:
        online:
          type: boolean
        players:
          type: integer
        maxPlayers:
          type: integer
        ping:
          type: integer
        uptime:
          type: string
        region:
          type: string
        lastUpdated:
          type: string

    NewsArticle:
      type: object
      required: [id, title, summary, content, author, publishedAt, category, imageUrl]
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        author:
          type: string
        publishedAt:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean

    GameRelease:
      type: object
      required: [id, version, releasedAt, isLatest, changelog]
      properties:
        id:
          type: integer
        version:
          type: string
        releasedAt:
          type: string
        isLatest:
          type: boolean
        changelog:
          type: array
          items:
            type: string
        downloadUrlAndroid:
          type: string
        downloadUrlWindows:
          type: string
        downloadUrlIos:
          type: string
        sizeAndroid:
          type: string
        sizeWindows:
          type: string
        sizeIos:
          type: string

    WikiEntry:
      type: object
      required: [id, slug, name, category, description]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        category:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        attributes:
          type: object
          additionalProperties:
            type: string
        lore:
          type: string

    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    RecoverInput:
      type: object
      required: [email]
      properties:
        email:
          type: string

    AuthResult:
      type: object
      required: [token, username]
      properties:
        token:
          type: string
        username:
          type: string

    MessageResult:
      type: object
      required: [message]
      properties:
        message:
          type: string