# lib/api-zod/src/generated/types/gameRelease.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sakura Chronicles MMORPG API
 * OpenAPI spec version: 0.1.0
 */

export interface GameRelease {
  id: number;
  version: string;
  releasedAt: string;
  isLatest: boolean;
  changelog: string[];
  downloadUrlAndroid?: string;
  downloadUrlWindows?: string;
  downloadUrlIos?: string;
  sizeAndroid?: string;
  sizeWindows?: string;
  sizeIos?: string;
}