# lib/api-zod/src/generated/types/serverStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sakura Chronicles MMORPG API
 * OpenAPI spec version: 0.1.0
 */

export interface ServerStatus {
  online: boolean;
  players: number;
  maxPlayers?: number;
  ping: number;
  uptime: string;
  region: string;
  lastUpdated: string;
}