# lib/api-zod/src/generated/types/wikiEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sakura Chronicles MMORPG API
 * OpenAPI spec version: 0.1.0
 */
import type { WikiEntryAttributes } from './wikiEntryAttributes';

export interface WikiEntry {
  id: number;
  slug: string;
  name: string;
  category: string;
  description: string;
  imageUrl?: string;
  attributes?: WikiEntryAttributes;
  lore?: string;
}